# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Generate New Ideas and List Existing Ones",
        "title": "Idea Generator",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/idea": {
            "get": {
                "description": "Returns the Data of All the Ideas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ideas"
                ],
                "summary": "Get a list of all ideas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Idea"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "description": "Creates New Data of a New Idea",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ideas"
                ],
                "summary": "Create Ideas",
                "parameters": [
                    {
                        "description": "Idea Name and Owner please",
                        "name": "EnterNameOwner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Idea"
                        }
                    }
                }
            }
        },
        "/idea/{id}": {
            "get": {
                "description": "Returns the Data of an Idea By it's ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ideas"
                ],
                "summary": "Get An Idea by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "idea ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Idea"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "description": "Updates Data of an Existing Idea",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ideas"
                ],
                "summary": "Update Ideas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "idea ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Idea Name and Owner please",
                        "name": "EnterNameOwner",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Idea"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "description": "Deletes the Data of an Idea By it's ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ideas"
                ],
                "summary": "Delete an Idea",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "idea ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Idea"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Idea": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Id is the primary key of the idea, and it is marked with the \"primary_key\" tag.",
                    "type": "integer"
                },
                "idea": {
                    "description": "IdeaName represents the name/title of the idea.",
                    "type": "string"
                },
                "owner": {
                    "description": "IdeaOwner represents the owner/creator of the idea.",
                    "type": "string"
                }
            }
        }
    }
}